requirements.txt
scrapy
airtable-python-wrapper
selenium
scrapy-selenium